Move agent sockets logic into its own file
